PdfDestination: don't create unused objects
As far as I can tell the only reason this was being done is so that the
array object has somewhere to live before it's copied with
AddToDictionary.

Since we have to add a PdfObject for this anyway, we can use it to
replace the PdfArray and PdfObject* fields.

The main caveat (and downside) of this patch is that when you initialize
the class with an existing object, it does copying during
initialziation. This makes GetObject pretty much useless.